Fix delete and move so they work on balls

+++ src/gizmoball/controller/DeleteGizmoListener.java
@@ -20,8 +20,8 @@ public class DeleteGizmoListener implements MouseListener {
     
     @Override
     public void mouseClicked(MouseEvent e) {
-        double chosenX = e.getX()/BoardView.L_TO_PIXELS;
-        double chosenY = e.getY()/BoardView.L_TO_PIXELS;
+        double chosenX = e.getX()/(double)BoardView.L_TO_PIXELS;
+        double chosenY = e.getY()/(double)BoardView.L_TO_PIXELS;
         
         model.select(chosenX, chosenY);
         model.delete();
