+++ src/gizmoball/controller/MoveGizmoListener.java
@@ -23,8 +23,8 @@ public class MoveGizmoListener implements MouseListener {
     
     @Override
     public void mouseClicked(MouseEvent e) {
-        double chosenX = e.getX()/BoardView.L_TO_PIXELS;
-        double chosenY = e.getY()/BoardView.L_TO_PIXELS;
+        double chosenX = e.getX()/(double)BoardView.L_TO_PIXELS;
+        double chosenY = e.getY()/(double)BoardView.L_TO_PIXELS;
         if (componentSelected) {
             try {
                 model.move(chosenX, chosenY);
